Login y crud de empleados OK

+++ src/main/java/com/Pro_Clinica_Veterinaria/demo/repository/EmpleadoRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface EmpleadoRepository extends PagingAndSortingRepository<Empleado, Long> {
+public interface EmpleadoRepository extends PagingAndSortingRepository<Empleado, Integer> {
 
 }
